+++ mocks/emailPairs-event.json
@@ -1,0 +1,3 @@
+{
+    "body": "[[{\"organisationName\":\"ABC Test Corp\",\"email\":\"mikkel.hlauritzen@gmail.com\",\"firstName\":\"Mikkel\"}, {\"organisationName\":\"ABC Test Corp\",\"email\":\"mikkel1995@live.dk\",\"firstName\":\"Makkk\"}],[{\"organisationName\":\"ABC Test Corp\",\"email\":\"mikkel.hlauritzen@gmail.com\",\"firstName\":\"Mikkel\"}, {\"organisationName\":\"ABC Test Corp\",\"email\":\"mikkel1995@live.dk\",\"firstName\":\"Makkk\"}]]"
+}
